move conversions to point of use to eliminate dep magnet
